+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   the line
 * Fixed `Screen::cursor_state_formatted` to draw the last character in a line
   with the appropriate drawing attributes if it needs to redraw it
+* Maintain wrapped state when overwriting a wide character at the end of a line
 
 ## [0.13.0] - 2021-11-17